Add update_scripts rake task.
They don't need to recompile if the only changes are to the scripts. One of the advantages of using an interpreted language :)
Idea: when they generatea new class or in any other way use our tooling to change something in the java, note that the app is "dirty" and needs to be compiled for the what's on the phone to reflect their current. Like in a .ruboto file or the like. Obviously won't help if they manually change stuff.