Added :margins parameter to View
In the Android SDK, margins are set within the layout holding the view, along with height & width.

In Ruboto, a View can accept the keywords height & width directly (as well as specifying them in the layout block for the view).

For consistency, allow the View to also accept the margins keyword.

All parameters that can be passed to the layout block are unaffected.

Example:

text_view :margins => [100,50,0,0]

is equivalent to

text_view :layout => {:set_margins => [100,50,0,0]}

which is equivalent to

text_view :layout => {:left_margin= => 100, :right_margin= => 50}