Added LayoutParams::MATCH_PARENT constant to view
Since Android 8, FILL_PARENT is deprecated in favor of MATCH_PARENT

Rather than defining each constant individually, use reflection on the Java class to get all constants and then register with View.

This ensures if a build is against an Android version earlier than 8 the Ruboto classes do not break the build.  If MATCH_PARENT was explicitly defined then we would  need to check the android version before defining, otherwise there would be a Constant not found error.

With reflection we define only the constants that are available to the current Android build.  In this situation the burden falls to the developer to only use constants that are present in the builds they are targeting.